+++ test/SemaphorePrio.cpp
@@ -52,11 +52,11 @@ TEST(SemaphorePrio) {
 		}
 	} t3(sem), t4(sem), t5(sem);
 
-	t1.start(0);
-	t2.start(0);
-	t3.start(1);
-	t4.start(1);
-	t5.start(1);
+	t1.start(uint8_t(0));
+	t2.start(uint8_t(0));
+	t3.start(uint8_t(1));
+	t4.start(uint8_t(1));
+	t5.start(uint8_t(1));
 	sem.init(false);
 
 	CommonTestUtils::start<OsRrPrio>();
